requirements.txt
python-dotenv==0.15.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.1
Django>=3.0,<4.0
psycopg2>=2.8
Pillow>=9.0
